layout:blog banner autoplay

+++ src/components/Blog/BlogDetail/CommentList/Reply/index.js
@@ -1,19 +1,58 @@
 import React from 'react'
 import moment from 'moment'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { useSelector } from 'react-redux'
 import { useDeleteReplyMutation } from '../../../../../services/replyApi'
+import Swal from 'sweetalert2'
 
 const Reply = ({ id, avatar, createTime, name, reply }) => {
   const [deleteReply] = useDeleteReplyMutation()
 
-  //想個辦法同步執行
-  const handleDelete = async (e) => {
-    const parentEel = e.target.parentElement
-    const targetId = parentEel.dataset.id
-    if (targetId === undefined) return
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger',
+    },
+    buttonsStyling: false,
+  })
 
-    await deleteReply({ id: targetId })
+  const handleDelete = async (e) => {
+    try {
+      const parentEel = e.target.parentElement
+      const targetId = parentEel.dataset.id
+      if (targetId === undefined) return
+
+      await swalWithBootstrapButtons
+        .fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonText: 'Yes, delete it!',
+          cancelButtonText: 'No, cancel!',
+          reverseButtons: true,
+        })
+        .then((result) => {
+          if (result.isConfirmed) {
+            swalWithBootstrapButtons.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+            deleteReply({ id: targetId })
+          } else if (
+            /* Read more about handling dismissals below */
+            result.dismiss === Swal.DismissReason.cancel
+          ) {
+            swalWithBootstrapButtons.fire(
+              'Cancelled',
+              'Your imaginary file is safe :)',
+              'error'
+            )
+          }
+        })
+    } catch (e) {
+      console.log(e)
+    }
   }
 
   return (
